Cleaned up lib structure, fixed tests
