--- lib/lib.php
@@ -1,7 +1,0 @@
-<?php
-
-// Entry point to the CRT library. This file should be preloaded.
-
-// Import all of the library modules
-require_once('crt.php');
-require_once('io.php');
