+++ lib/tests/ErrorTest.php
@@ -1,7 +1,5 @@
 <?php
 
-require_once('tests.inc');
-
 final class ErrorTest extends PHPUnit_Framework_TestCase {
     private static $crt = null;
     public static function setUpBeforeClass() {
